requirements.txt
google-generativeai
python-dotenv
flask
gunicorn
pandas
numpy
matplotlib
python-dateutil
scikit-learn